requirements.txt
fastapi
uvicorn
httpx
google-generativeai
python-dotenv
pyzbar
google-cloud-vision
